Extract tap_init's logic into InitializationTracker class
