+++ lib/tapping_device/trackable.rb
@@ -1,9 +1,13 @@
 class TappingDevice
   module Trackable
-    [:tap_on!, :tap_init!, :tap_assoc!, :tap_passed!].each do |method|
+    [:tap_on!, :tap_assoc!, :tap_passed!].each do |method|
       define_method method do |object, options = {}, &block|
         new_device(options, &block).send(method, object)
       end
+    end
+
+    def tap_init!(object, options = {}, &block)
+      TappingDevice::Trackers::InitializationTracker.new(options, &block).start_tracking(object)
     end
 
     def print_traces(target, options = {})
